==> 1_BookValue.ipynb <==
import pandas as pd
import numpy as np
from pandas import DataFrame

# Read data
data = pd.read_csv('RawData/FirmData.csv')
data["Year"] = pd.to_datetime(data["Date"], dayfirst=True).dt.year
data.head(5)

# Convert dataframe to correct format
df_Book_Value = data.pivot(index="RIC", columns="Year", values="Total Equity")
for i in df_Book_Value.columns:
    df_Book_Value[i] = pd.to_numeric(df_Book_Value[i], errors='coerce')
df_Book_Value.head(5)

# Check if all columns have data type float64 e.g. only contain numbers
df_Book_Value.dtypes

# Write dataframe
df_Book_Value.to_pickle("UsableData/BookValue.pkl")
