==> 1_EligibleStocks.ipynb <==
import pandas as pd
import numpy as np
from pandas import DataFrame
import math

# Read dataframes
df_Market_Cap = pd.read_pickle("UsableData/MarketCap.pkl")
df_Employees = pd.read_pickle("UsableData/Employees.pkl")
df_Book_Value = pd.read_pickle("UsableData/BookValue.pkl")
df_Variance = pd.read_pickle("UsableData/Variance.pkl")
df_Sales = pd.read_pickle("UsableData/Sales.pkl")
df_Returns = pd.read_pickle("UsableData/Returns.pkl").T

# Use existing dataframe as template and set all cloumns to False and data type to Boolean
df_eligible = df_Market_Cap.copy(deep=True)
for column in df_eligible:
    df_eligible[column].astype("bool")
    df_eligible[column] = False

# Create method that checks if there are non NaN returns in one year and therefore the stock is usable in that year in regards to the return data
def unusableReturns(ric, year):
    selection = df_Returns[ric][(df_Returns.index >= f"{year}-01-01") & (df_Returns.index <= f"{year}-12-31")]
    return bool(selection.isna().values.any())

# Iterate through every year and see which stocks are eligible
years = df_Market_Cap.columns.to_numpy()[:-1] # Exclude first year, because no data from previous year to calculate weights
for ric in df_Market_Cap.index: # Go through every stock
    for year in years: # In every year
        mc = df_Market_Cap.loc[ric, year]
        emp = df_Employees.loc[ric, year]
        bv = df_Book_Value.loc[ric, year]
        var = df_Variance.loc[ric, year]
        sal = df_Sales.loc[ric, year]
        if(math.isnan(mc) or math.isnan(emp) or math.isnan(bv) or math.isnan(var) or math.isnan(sal) or unusableReturns(ric, year+1)): # If one of the data for weights (mc, emp, bv, var) is nan or return for next year stock is not eligible
            df_eligible.at[ric, year+1] = False
        else: # If all the data is there, stock is eligible
            df_eligible.at[ric, year+1] = True
df_eligible.drop(columns=[years[0]], inplace=True) # Drop first and last year
df_eligible.head(5)

# Count how many stocks are eligible in each year
for i in range(len(df_eligible.columns)):
    df = df_eligible[df_eligible[df_eligible.columns[i]] == True]
    print(f"{df_eligible.columns[i]}: {len(df.index)}")

# Write dataframe
df_eligible.to_pickle("UsableData/EligibleStocks.pkl")
