==> 1_EqualRandom.ipynb <==
import pandas as pd
import numpy as np

# Read existing dataframe as template
df_Market_Cap = pd.read_pickle("UsableData/MarketCap.pkl")

# Fill dataframe with ones for equal weighting
df_Equal = df_Market_Cap.copy(deep=True)
for column in df_Equal:
    df_Equal[column] = 1
df_Equal.to_pickle("UsableData/Equal.pkl")
df_Equal.head(5)

# Fill dataframe with random numbers for random weighting
df_Random = df_Market_Cap.copy(deep=True)
for column in df_Random.columns:
    df_Random[column] = np.random.random([len(df_Random.index)])
df_Random.to_pickle("UsableData/Random.pkl")
df_Random.head(5)
