==> 1_Sales.ipynb <==
import pandas as pd
import numpy as np
from pandas import DataFrame
from sklearn.preprocessing import minmax_scale

# Read data
data = pd.read_csv('RawData/FirmData.csv')
data["Year"] = pd.to_datetime(data["Date"], dayfirst=True).dt.year
data.head(5)

# Convert dataframe to correct format
df_Sales = data.pivot(index="RIC", columns="Year", values="Revenue")
for i in df_Sales.columns:
    df_Sales[i] = pd.to_numeric(df_Sales[i], errors='coerce')
df_Sales.head(5)

for year in df_Sales.columns:
    df_Sales[year] = minmax_scale(df_Sales[year])
df_Sales

# Check if all columns have data type float64 e.g. only contain numbers
df_Sales.dtypes

# Write dataframe
df_Sales.to_pickle("UsableData/Sales.pkl")
