==> 1_Variance.ipynb <==
import pandas as pd
import numpy as np
from pandas import DataFrame

# Read returns to calculate the variance of the returns and create empty dataframe
df_Returns = pd.read_pickle("UsableData/Returns.pkl").T
df_Variance = DataFrame()

# Calculate variance of returns for each year and stock
years = df_Returns.index.year.unique()
for year in years:
    returns = df_Returns[(df_Returns.index >= f"{year}-01-01") & (df_Returns.index <= f"{year}-12-31")]
    for ric in returns.columns:
        df_Variance.at[ric, year] = np.var(returns[ric])
df_Variance.head(5)

# Exclude zeros as this will result in infinite weight when calculating 1/var of that stock
df_Variance = df_Variance[~(df_Variance[df_Variance.columns] == 0)]

# Save dataframe
df_Variance.to_pickle("UsableData/Variance.pkl")
