==> 3_Analysis.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from scipy.stats import gmean
from pandas import DataFrame


df_Results = pd.read_pickle("UsableData/Results.pkl")
# df_Results = df_Results[df_Results.index < "2020"]
df_Results.tail(5)

# Based on 3.1 Finance Introduction
total_cumulative_return = (df_Results.iloc[-1]-df_Results.iloc[1])*100
total_cumulative_return

# Based on 3.1 Finance Introduction
mean_yearly_return = (((1+(total_cumulative_return/100)))**(1/10)-1)*100
mean_yearly_return

df_Returns_Strat = df_Results.pct_change(fill_method=None)*100
df_Returns_Strat.dropna(axis=0, inplace=True)
df_Returns_Strat.head(5)

monthly_vol = np.std(df_Returns_Strat, axis=0)
monthly_vol

# Read yearly return of DE 10 Year Bound reported monthly
DE10Y = pd.read_csv("RawData/DE10Y.csv", index_col="Datum")
DE10Y.index = pd.to_datetime(DE10Y.index, dayfirst=True)
DE10Y = DE10Y[(DE10Y.index >= df_Results.index[1]) & (DE10Y.index <= df_Results.index[-1])]
for row in DE10Y.index:
    string = DE10Y.loc[row, "Zuletzt"].strip("%").replace(".", "").replace(",", ".")
    DE10Y.at[row, "Zuletzt"] = float(string)
DE10Y.head(5)

# Based on 3.1 Finance Introduction
risk_free_return = np.mean(DE10Y["Zuletzt"])
risk_free_return

# Based on 3.1 Finance Introduction
sharpe_ratio = pd.Series(index=df_Results.columns)
for strategy in mean_yearly_return.index:
    sharpe_ratio[strategy] = (mean_yearly_return[strategy] - risk_free_return) / monthly_vol[strategy]
sharpe_ratio

# maximum monthly drawdown
max_monthly_drawdown = pd.Series(index=df_Results.columns)
for strategy in df_Returns_Strat.columns:
    max_monthly_drawdown[strategy] = min(df_Returns_Strat[strategy])
max_monthly_drawdown

positive_month = pd.Series(index=df_Results.columns)
for strategy in df_Returns_Strat.columns:
    returns = df_Returns_Strat[strategy]
    returns = returns[returns > 0]
    positive_month[strategy] = len(returns)
positive_month

df_Analysis = DataFrame(index=df_Results.columns)
df_Analysis["Total Cumulative Return"] = total_cumulative_return
df_Analysis["Mean Yearly Return"] = mean_yearly_return
df_Analysis["Volatility"] = monthly_vol
df_Analysis["Sharpe Ratio"] = sharpe_ratio
df_Analysis["Max Drawdown"] = max_monthly_drawdown
df_Analysis["Positive Month"] = positive_month
df_Analysis.to_pickle("UsableData/Analysis.pkl")
df_Analysis
